test(OrderTest): 테스트코드 클래스 분리
