+++ src/test/java/kitchenpos/eatinorders/tobe/domain/DeliveryOrderTest.java
@@ -1,0 +1,19 @@
+package kitchenpos.eatinorders.tobe.domain;
+
+/*
+- 1개 이상의 등록된 메뉴로 배달 주문을 등록할 수 있다.
+- 배달 주문을 접수되면 배달 대행사를 호출한다.
+- 주문을 배달한다.
+- 배달 주문만 배달할 수 있다.
+- 서빙된 주문만 배달할 수 있다.
+- 주문을 배달 완료한다.
+- 배달 중인 주문만 배달 완료할 수 있다.
+- 배달 주문의 경우 배달 완료된 주문만 완료할 수 있다.
+- 배달 주소가 올바르지 않으면 배달 주문을 등록할 수 없다.
+  - 배달 주소는 비워 둘 수 없다.
+- 배달 주문의 경우 주문 항목의 수량은 0 이상이어야 한다.
+ */
+class DeliveryOrderTest {
+
+
+}
